+++ src/main/java/org/summerboot/jexpress/nio/server/ws/rs/BootController.java
@@ -53,6 +53,7 @@ import org.summerboot.jexpress.boot.BootConstant;
 import org.summerboot.jexpress.boot.BootErrorCode;
 import org.summerboot.jexpress.boot.annotation.Daemon;
 import org.summerboot.jexpress.boot.annotation.Log;
+import org.summerboot.jexpress.boot.annotation.RequiresHealthCheck;
 import org.summerboot.jexpress.boot.instrumentation.HealthMonitor;
 import org.summerboot.jexpress.integration.cache.AuthTokenCache;
 import org.summerboot.jexpress.nio.server.SessionContext;
@@ -165,6 +166,7 @@ abstract public class BootController extends PingController {
     @Produces(MediaType.TEXT_HTML)
     @RolesAllowed({Config.ROLE_ADMIN})
     @Daemon
+    @RequiresHealthCheck("")
     //@CaptureTransaction("admin.version")
     public void version(@Parameter(hidden = true) final SessionContext context) {
         context.response(getVersion()).status(HttpResponseStatus.OK);
@@ -200,6 +202,7 @@ abstract public class BootController extends PingController {
     @Produces(MediaType.TEXT_HTML)
     @RolesAllowed({Config.ROLE_ADMIN})
     @Daemon
+    @RequiresHealthCheck("")
     //@CaptureTransaction("admin.inspect")
     public void inspect(@Parameter(hidden = true) final SessionContext context) {
         HealthMonitor.inspect();
@@ -225,6 +228,7 @@ abstract public class BootController extends PingController {
     @Path(Config.CURRENT_VERSION + Config.API_ADMIN_STATUS)
     @RolesAllowed({Config.ROLE_ADMIN})
     @Daemon
+    @RequiresHealthCheck("")
     //@CaptureTransaction("admin.changeStatus")
     public void pause(@QueryParam("pause") boolean pause, @Parameter(hidden = true) final SessionContext context) throws IOException {
         HealthMonitor.pauseService(pause, BootConstant.PAUSE_LOCK_CODE_VIAWEB, "request by " + context.caller());
@@ -254,6 +258,7 @@ abstract public class BootController extends PingController {
     @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
     @Path(Config.CURRENT_VERSION + Config.API_NF_JSECURITYCHECK)
     @Daemon
+    @RequiresHealthCheck("")
     //@CaptureTransaction("user.signJWT")
     @Log(requestBody = false, maskDataFields = Config.X_AUTH_TOKEN)
     public Caller longin_jSecurityCheck(@Parameter(required = true) @Nonnull @FormParam("j_username") String userId,
@@ -285,6 +290,7 @@ abstract public class BootController extends PingController {
     @Consumes(MediaType.APPLICATION_JSON)
     @Path(Config.CURRENT_VERSION + Config.API_NF_LOGIN)
     @Daemon
+    @RequiresHealthCheck("")
     //@CaptureTransaction("user.signJWT")
     @Log(requestBody = false, maskDataFields = Config.X_AUTH_TOKEN)
     public Caller longin_JSON(@Valid @Nonnull LoginVo loginVo,
@@ -336,6 +342,7 @@ abstract public class BootController extends PingController {
     @DELETE
     @Path(Config.CURRENT_VERSION + Config.API_NF_LOGIN)
     @Daemon
+    @RequiresHealthCheck("")
     //@PermitAll
     //@CaptureTransaction("user.logoutToken")
     public void logout(@Parameter(hidden = true) final ServiceRequest request, @Parameter(hidden = true) final SessionContext context) {
@@ -354,6 +361,7 @@ abstract public class BootController extends PingController {
     @Path(Config.CURRENT_VERSION + Config.API_NF_LOADTEST)// .../loadtest?delayMilsec=123
     @RolesAllowed({Config.ROLE_ADMIN})
     @Daemon
+    @RequiresHealthCheck("")
     public void loadTestBenchmarkPost1(final ServiceRequest request, final SessionContext context, @QueryParam("delayMilsec") long wait) {
         if (wait > 0) {
             try {
@@ -369,6 +377,7 @@ abstract public class BootController extends PingController {
     @POST
     @Path(Config.CURRENT_VERSION + Config.API_NF_LOADTEST + "/{delayMilsec}")
     @Daemon
+    @RequiresHealthCheck("")
     public void loadTestBenchmarkPost2(final ServiceRequest request, final SessionContext context, @PathParam("delayMilsec") long wait) {
         if (wait > 0) {
             try {
@@ -385,6 +394,7 @@ abstract public class BootController extends PingController {
     @Path(Config.CURRENT_VERSION + Config.API_NF_LOADTEST)// .../loadtest?delayMilsec=123
     @RolesAllowed({Config.ROLE_ADMIN})
     @Daemon
+    @RequiresHealthCheck("")
     public void loadTestBenchmarkGet1(final ServiceRequest request, final SessionContext context, @QueryParam("delayMilsec") long wait) {
         if (wait > 0) {
             try {
@@ -400,6 +410,7 @@ abstract public class BootController extends PingController {
     @GET
     @Path(Config.CURRENT_VERSION + Config.API_NF_LOADTEST + "/{delayMilsec}")
     @Daemon
+    @RequiresHealthCheck("")
     public void loadTestBenchmarkGet2(final ServiceRequest request, final SessionContext context, @PathParam("delayMilsec") long wait) {
         if (wait > 0) {
             try {